put db connection in module
