+++ index.js
@@ -99,6 +99,7 @@ const CUSTOM_HANDLERS = [
     'untagged',
     'no-icon',
     'no-image',
+    'missing-image',
     'no-wiki',
     'all',
 ];
@@ -173,20 +174,26 @@ const getTableContents = async (filterObject) => {
                     }
                     break;
                 case 'no-image':
-                    if(item.image_link !== ''){
+                    if(item.image_link){
                         continue;
                     }
 
                     break;
                 case 'no-icon':
-                    if(item.icon_link !== ''){
+                    if(item.icon_link){
                         continue;
                     }
 
                     break;
+                case 'missing-image':
+                    if(item.image_link && item.icon_link && item.grid_image_link){
+                        continue;
+                    }
 
+                    break;
+    
                 case 'no-wiki':
-                    if(item.wiki_link !== ''){
+                    if(item.wiki_link){
                         continue;
                     }
