+++ jobs/clear-checkouts.js
@@ -1,13 +1,4 @@
-const mysql = require('mysql');
-
-const connection = mysql.createConnection({
-    host     : 'tarkov-tools-master-1.cluster-c1vhfeufwkpn.eu-west-1.rds.amazonaws.com',
-    user     : 'desktop1',
-    password : process.env.MYSQL_PASSWORD,
-    database : 'tarkov_tools',
-});
-
-connection.connect();
+const connection = require('../modules/db-connection');
 
 module.exports = async () => {
     const promise = new Promise((resolve, reject) => {
