+++ modules/do-query.js
@@ -1,15 +1,6 @@
-const mysql = require('mysql');
+const connection = require('./db-connection');
 
 module.exports = async function doQuery(query, params) {
-    const connection = mysql.createConnection({
-        host     : 'tarkov-tools-master-1.cluster-c1vhfeufwkpn.eu-west-1.rds.amazonaws.com',
-        user     : 'desktop1',
-        password : process.env.MYSQL_PASSWORD,
-        database : 'tarkov_tools',
-    });
-
-    connection.connect();
-
     let responseData;
     const promise = new Promise((resolve, reject) => {
         connection.query(query,
